Refactored file transfer to get around error

+++ sample-clockface/app/logFileTransfer.ts
@@ -5,28 +5,37 @@ const LOG_FILE_NAME = "./resources/logFile.txt";
 
 export function exportLogs() {
     try {
-            console.log("Exporting test log file");
+        let promise = fileTransfer.canTransfer();
+        return promise.then(function (underLimit:Boolean) {
 
-            //Set a unique transfer file name so we don't overwrite any previously queued transfer files
-            let logTransferFileName = `${Date.now()}_logFile.txt`;
+            if (underLimit) {
+                console.log("Exporting test log file");
 
-            try {
-                let ab = fs.readFileSync(LOG_FILE_NAME);
-                fs.writeFileSync(logTransferFileName, ab);
-            } catch(e) {
-                console.log("Error renaming log file");
-                console.log(e);
-                return;
-            }
+                //Set a unique transfer file name so we don't overwrite any previously queued transfer files
+                let logTransferFileName = `${Date.now()}_logFile.txt`;
 
-            fileTransfer.transferFile(logTransferFileName)
-            .then(function() {
-                console.log("Export complete");
+                try {
+                    let ab = fs.readFileSync(LOG_FILE_NAME);
+                    fs.writeFileSync(logTransferFileName, ab);
+                } catch(e) {
+                    console.log("Error renaming log file");
+                    console.log(e);
+                    return;
+                }
+
+                fileTransfer.queueFileTransfer(logTransferFileName)
+                .then(function() {
+                    console.log("Export complete");
+                    fs.unlinkSync(logTransferFileName);
+                });
+                
                 fs.unlinkSync(logTransferFileName);
-            });
-            
-            fs.unlinkSync(logTransferFileName);
-        
+            } else {
+                let error = new Error("Cannot queue file transfer, max transfer count reached");
+                console.log(error);
+                return Promise.reject(error);
+            }
+        });
     } catch(ex) {
         console.log(ex);
     }
